pyproject.toml
[tool.poetry]
name = "bareunpy"
version = "1.4.0"
description = "The bareun python library using grpc"
authors = ["Gihyun YUN <gih2yun@baikal.ai>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://bareun.ai/"
repository = "https://github.com/bareun-nlp/bareunpy"
keywords = [ "NLP", "Korean", "Deep Learning", "POS tagger", "bareun"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Education",
    "Natural Language :: Korean",
    "Operating System :: OS Independent",
    "Typing :: Typed",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Operating System :: Unix",
    "Operating System :: MacOS"
]

[tool.poetry.dependencies]
python = "^3.6"
grpcio = "^1.46.0"
protobuf = "^3.19.4"
googleapis-common-protos = "^1.56.0"
bareun-apis = "^0.10.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
